uploaded with docker image

+++ services/user-service/src/main/java/com/assesment/userservice/repository/UserDetailsRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserDetailsRepo extends JpaRepository<UserDetails,Long> {
+public interface UserDetailsRepo extends JpaRepository<UserDetails,Integer> {
 
-    UserDetails findAllById(Long id);
+    UserDetails findAllById(Integer id);
 }
